move to own crate + print msg
