+++ zkvm/entrypoint/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 sp1-precompiles = { path = "../precompiles" }
-anyhow = "1.0.75"
 bincode = "1.3.3"
 cfg-if = "1.0.0"
 getrandom = { version = "0.2.12", features = ["custom"] }
